creates separate file for messages
